# pubquiz/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API",
        "description": "API documentation for Pub Quiz application.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/seasons": {
            "get": {
                "tags": [
                    "Seasons"
                ],
                "summary": "List all seasons",
                "operationId": "3265d3d787710ae89d8c32fe652c9968",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "seasons": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Season 2025"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "season-2025"
                                                    },
                                                    "start_date": {
                                                        "type": "string",
                                                        "example": "2025-01-15"
                                                    },
                                                    "end_date": {
                                                        "type": "string",
                                                        "example": "2025-06-15"
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Spring/Summer season."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No seasons found."
                    }
                }
            },
            "post": {
                "tags": [
                    "Seasons"
                ],
                "summary": "Create a season (admin only)",
                "operationId": "5535d31c60ce7987e77ee05ddf7731d9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Season 2026"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "season-2026"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "example": "2026-02-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "example": "2026-06-01"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Early-year season."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Season created successfully"
                                        },
                                        "season": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Season 2026"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "season-2026"
                                                },
                                                "start_date": {
                                                    "type": "string",
                                                    "example": "2026-02-01"
                                                },
                                                "end_date": {
                                                    "type": "string",
                                                    "example": "2026-06-01"
                                                },
                                                "is_active": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Early-year season."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Only admins can create seasons"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/seasons/{id}/board": {
            "get": {
                "tags": [
                    "Seasons"
                ],
                "summary": "Season TOTAL board and per-event boards",
                "operationId": "cac4470092141820b3c2e58bc41cae61",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "season": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Season 2025"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "total_board": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "team": {
                                                        "type": "string",
                                                        "example": "Quiz Masters"
                                                    },
                                                    "points": {
                                                        "type": "integer",
                                                        "example": 245
                                                    },
                                                    "rank": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "events": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "event": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 10
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Belgrade Pub Quiz #1"
                                                            },
                                                            "starts_at": {
                                                                "type": "string",
                                                                "example": "2025-03-10T19:00:00"
                                                            },
                                                            "status": {
                                                                "type": "string",
                                                                "example": "completed"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "board": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "team": {
                                                                    "type": "string",
                                                                    "example": "Quiz Masters"
                                                                },
                                                                "points": {
                                                                    "type": "integer",
                                                                    "example": 87
                                                                },
                                                                "rank": {
                                                                    "type": "integer",
                                                                    "example": 1,
                                                                    "nullable": true
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No events / participations found."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}